requirements.txt
python_telegram_bot==13.7
yt_dlp
